StyleSheet refactor
