+++ src/StyleSheet/generateCss.js
@@ -1,11 +1,7 @@
-const prefixAll = require('inline-style-prefix-all');
-const processTransform = require('./processTransform');
 const normalizeValue = require('./normalizeValue');
-const expandStyle = require('./expandStyle');
 const hyphenate = require('./hyphenate');
 const mapKeyValue = require('../util/mapKeyValue');
-
-const normalizeStyle = style => prefixAll(processTransform(expandStyle(style)));
+const transformToWebStyle = require('./transformToWebStyle');
 
 /**
  * Generates valid CSS rule bodies (sans selector) from a JS object
@@ -18,9 +14,12 @@ const normalizeStyle = style => prefixAll(processTransform(expandStyle(style)));
  *
  * @param style
  */
-const generateCss = style => mapKeyValue(normalizeStyle(style), (key, val) => {
+const generateCss = style => mapKeyValue(transformToWebStyle(style), (key, val) => {
   const name = hyphenate(key);
   const value = normalizeValue(key, val);
+  if (Array.isArray(val)) {
+    return val.map(v => `${name}:${v};`).join('');
+  }
   return `${name}:${value};`;
 }).join('');
