+++ src/StyleSheet/getHairlineWidth.js
@@ -1,11 +1,13 @@
 // Modeled after: http://dieulot.net/css-retina-hairline
 
-// TODO(lmr): come up with an SSR strategy for this...
-
 // NOTE:
-// This assumes this script runs after <body>. If it is run inside of `<head>`, it would not work
+// This assumes this script runs after <body>. If it is run inside of `<head>`, it will just
+// return 1.
 module.exports = () => {
   let hairlineWidth = 1;
+  if (!global || !global.document || !global.document.body) {
+    return hairlineWidth;
+  }
   if (global.devicePixelRatio && global.devicePixelRatio >= 2) {
     const testElem = document.createElement('div');
     testElem.style.border = '.5px solid transparent';
