+++ src/StyleSheet/prefixedProperties.js
@@ -1,0 +1,138 @@
+module.exports = ([
+  // 'animation',
+  // 'animationDelay',
+  // 'animationDirection',
+  // 'animationFillMode',
+  // 'animationDuration',
+  // 'animationIterationCount',
+  // 'animationName',
+  // 'animationPlayState',
+  // 'animationTimingFunction',
+  // 'appearance',
+  // 'backdropFilter',
+  // 'backfaceVisibility',
+  'perspective',
+  'perspectiveOrigin',
+  'transform',
+  'transformOrigin',
+  'transformStyle',
+  'transformOriginX',
+  'transformOriginY',
+  'transformOriginZ',
+  'backgroundClip',
+  'backgroundOrigin',
+  'backgroundSize',
+  // 'borderImage',
+  // 'borderImageOutset',
+  // 'borderImageRepeat',
+  // 'borderImageSlice',
+  // 'borderImageSource',
+  // 'borderImageWidth',
+  'borderRadius',
+  // 'boxDecorationBreak',
+  'boxShadow',
+  'boxSizing',
+  // 'breakAfter',
+  // 'breakBefore',
+  // 'breakInside',
+  // 'columnCount',
+  // 'columnFill',
+  // 'columnGap',
+  // 'columnRule',
+  // 'columnRuleColor',
+  // 'columnRuleStyle',
+  // 'columnRuleWidth',
+  // 'columns',
+  // 'columnSpan',
+  // 'columnWidth',
+  // 'clipPath',
+  'filter',
+  'flex',
+  'flexBasis',
+  'flexDirection',
+  'flexGrow',
+  'flexFlow',
+  'flexShrink',
+  'flexWrap',
+  'alignContent',
+  'alignItems',
+  'alignSelf',
+  'justifyContent',
+  'order',
+  // 'flowInto',
+  // 'flowFrom',
+  // 'breakBefore',
+  // 'breakAfter',
+  // 'breakInside',
+  // 'regionFragment',
+  // 'fontFeatureSettings',
+  'fontKerning',
+  // 'gridTemplateColumns',
+  // 'gridTemplateRows',
+  // 'gridTemplateAreas',
+  // 'gridTemplate',
+  // 'gridAutoColumns',
+  // 'gridAutoRows',
+  // 'gridAutoFlow',
+  // 'grid',
+  // 'gridRowStart',
+  // 'gridColumnStart',
+  // 'gridRowEnd',
+  // 'gridRow',
+  // 'gridColumn',
+  // 'gridArea',
+  // 'rowGap',
+  // 'columnGap',
+  // 'gridGap',
+  // 'hyphens',
+  // 'maskImage',
+  // 'maskMode',
+  // 'maskRepeat',
+  // 'maskPosition',
+  // 'maskClip',
+  // 'maskOrigin',
+  // 'maskSize',
+  // 'maskComposite',
+  // 'mask',
+  // 'maskBorderSource',
+  // 'maskBorderMode',
+  // 'maskBorderSlice',
+  // 'maskBorderWidth',
+  // 'maskBorderOutset',
+  // 'maskBorderRepeat',
+  // 'maskBorder',
+  // 'maskType',
+  // 'objectFit',
+  // 'objectPosition',
+  // 'resize',
+  // 'scrollSnapType',
+  // 'scrollSnapPointsX',
+  // 'scrollSnapPointsY',
+  // 'scrollSnapDestination',
+  // 'scrollSnapCoordinate',
+  // 'shapeImageThreshold',
+  // 'shapeImageMargin',
+  // 'shapeImageOutside',
+  // 'tabSize',
+  'textAlignLast',
+  'textDecorationStyle',
+  'textDecorationSkip',
+  'textDecorationLine',
+  'textDecorationColor',
+  // 'textEmphasisPosition',
+  // 'textEmphasis',
+  // 'textEmphasisStyle',
+  // 'textEmphasisColor',
+  'textOverflow',
+  'textSizeAdjust',
+  // 'touchAction',
+  // 'transition',
+  // 'transitionDelay',
+  // 'transitionDuration',
+  // 'transitionProperty',
+  // 'transitionTimingFunction',
+  'userSelect',
+  // 'wrapFlow',
+  // 'wrapThrough',
+  // 'wrapMargin',
+]).reduce((acc, key) => { acc[key] = true; return acc; }, {});
